+++ Project/Assets/ML-Agents/Examples/PushBlock/Scripts/PushAgentBasic.cs
@@ -54,7 +54,7 @@ public class PushAgentBasic : Agent
     protected override void Awake()
     {
         base.Awake();
-        m_PushBlockSettings = FindObjectOfType<PushBlockSettings>();
+        m_PushBlockSettings = FindFirstObjectByType<PushBlockSettings>();
     }
 
     public override void Initialize()
@@ -205,7 +205,7 @@ public class PushAgentBasic : Agent
         block.transform.position = GetRandomSpawnPos();
 
         // Reset block velocity back to zero.
-        m_BlockRb.velocity = Vector3.zero;
+        m_BlockRb.linearVelocity = Vector3.zero;
 
         // Reset block angularVelocity back to zero.
         m_BlockRb.angularVelocity = Vector3.zero;
@@ -223,7 +223,7 @@ public class PushAgentBasic : Agent
 
         ResetBlock();
         transform.position = GetRandomSpawnPos();
-        m_AgentRb.velocity = Vector3.zero;
+        m_AgentRb.linearVelocity = Vector3.zero;
         m_AgentRb.angularVelocity = Vector3.zero;
 
         SetResetParameters();
@@ -244,7 +244,7 @@ public class PushAgentBasic : Agent
         m_BlockRb.transform.localScale = new Vector3(scale, 0.75f, scale);
 
         // Set the drag of the block
-        m_BlockRb.drag = m_ResetParams.GetWithDefault("block_drag", 0.5f);
+        m_BlockRb.linearDamping = m_ResetParams.GetWithDefault("block_drag", 0.5f);
     }
 
     void SetResetParameters()