+++ Project/Assets/ML-Agents/Examples/PushBlock/Scripts/PushAgentCollab.cs
@@ -12,7 +12,7 @@ public class PushAgentCollab : Agent
     protected override void Awake()
     {
         base.Awake();
-        m_PushBlockSettings = FindObjectOfType<PushBlockSettings>();
+        m_PushBlockSettings = FindFirstObjectByType<PushBlockSettings>();
     }
 
     public override void Initialize()