+++ Project/Assets/ML-Agents/Examples/PushBlockWithInput/Scripts/PushBlockWithInputPlayerController.cs
@@ -30,7 +30,7 @@ public class PushBlockWithInputPlayerController : MonoBehaviour, IInputActionAss
 
     void Awake()
     {
-        m_PushBlockSettings = FindObjectOfType<PushBlockWithInputSettings>();
+        m_PushBlockSettings = FindFirstObjectByType<PushBlockWithInputSettings>();
         LazyInitializeActions();
 
         // Cache the agent rigidbody